ours/requirements.txt
pydicom
numpy
nibabel
scipy
Pillow
imageio